+++ src/assets/images/logo-bg.png
