+++ src/components/header/Header.jsx
@@ -1,5 +1,5 @@
+import textLogoImg from '@/assets/images/logo-text-temp.png';
 import logoImg from '@/assets/images/logo.png';
-import textLogoImg from '@/assets/images/logo_text_temp.png';
 import profileImg from '@/assets/images/stupid.png';
 import { Link } from 'react-router-dom';
 import * as S from './header.styles';
