+++ src/components/modals/creditChargeModal/index.js
@@ -1,1 +1,1 @@
-export { default } from './creditChargeModal';
+export { default } from './CreditChargeModal';
